web/requirements.txt
Flask==0.12.2
Jinja2==2.9.6
python-google-places==1.4.0
